requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
streamlit>=1.28.0
gradio>=4.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
albumentations>=1.3.0
tensorboard>=2.13.0
